Improved UI and formated frontend code
