+++ frontend/.eslintrc.cjs
@@ -13,9 +13,6 @@ module.exports = {
     plugins: ["react-refresh"],
     rules: {
         "react/jsx-no-target-blank": "off",
-        "react-refresh/only-export-components": [
-            "warn",
-            { allowConstantExport: true },
-        ],
+        "react-refresh/only-export-components": ["warn", { allowConstantExport: true }],
     },
 }
