+++ frontend/src/App.jsx
@@ -1,15 +1,18 @@
-import { Routes, Route, Outlet, Navigate } from "react-router-dom"
+import { Navigate, Outlet, Route, Routes } from "react-router-dom"
 import { useSignalEffect, useSignals } from "@preact/signals-react/runtime"
-import { userExists, userDetails, userDetailsUpdate } from "@/signals/user"
+
+import { userDetails, userDetailsUpdate, userExists } from "@/signals/user"
+import axiosInstance from "@/axios"
+
+import Account from "@/pages/Account"
+import AuthGoogle from "@/pages/AuthGoogle"
+import Login from "@/pages/Login"
+import Read from "@/pages/Read"
+import Signup from "@/pages/Signup"
+import Footer from "@/components/Footer"
 import Hero from "@/components/Hero"
 import Navbar from "@/components/Navbar"
-import Footer from "@/components/Footer"
-import Login from "@/components/Login"
-import Signup from "@/components/Signup"
-import Read from "@/components/Read"
-import AuthGoogle from "@/components/AuthGoogle"
-import Account from "@/components/Account"
-import axiosInstance from "@/axios"
+
 import "./App.css"
 
 function Layout() {
@@ -54,18 +57,9 @@ function App() {
                     <>
                         <Route index element={<Read />} />
                         <Route path="/account" element={<Account />} />
-                        <Route
-                            path="read"
-                            element={<Navigate to="/" replace={true} />}
-                        />
-                        <Route
-                            path="login"
-                            element={<Navigate to="/read" replace={true} />}
-                        />
-                        <Route
-                            path="signup"
-                            element={<Navigate to="/read" replace={true} />}
-                        />
+                        <Route path="read" element={<Navigate to="/" replace={true} />} />
+                        <Route path="login" element={<Navigate to="/read" replace={true} />} />
+                        <Route path="signup" element={<Navigate to="/read" replace={true} />} />
                         <Route
                             path="auth/google"
                             element={<Navigate to="/" replace={true} />}
@@ -77,14 +71,8 @@ function App() {
                         <Route path="signup" element={<Signup />} />
                         <Route path="login" element={<Login />} />
                         <Route path="auth/google" element={<AuthGoogle />} />
-                        <Route
-                            path="read"
-                            element={<Navigate to="/login" replace={true} />}
-                        />
-                        <Route
-                            path="*"
-                            element={<Navigate to="/" replace={true} />}
-                        />
+                        <Route path="read" element={<Navigate to="/login" replace={true} />} />
+                        <Route path="*" element={<Navigate to="/" replace={true} />} />
                     </>
                 )}
             </Route>
