+++ frontend/src/axios.js
@@ -1,4 +1,5 @@
 import axios from "axios"
+
 import { getCookie, setCookie } from "./helpers/cookies"
 import { userExists } from "./signals/user"
