+++ frontend/src/components/ArticleCard.jsx
@@ -2,15 +2,15 @@ export const ArticleCard = ({ article }) => {
     return (
         <div className="group flex h-fit flex-col gap-3 hover:cursor-pointer">
             <img
-                className="h-46 aspect-[4/3] rounded-sm object-cover"
+                className="w-full aspect-[4/3] rounded-sm object-cover"
                 src={article.image}
                 alt=""
             />
             <div className="flex flex-col justify-center gap-2">
-                <span className="font-serif text-lg font-medium leading-6 group-hover:underline group-hover:underline-offset-4">
+                <span className="font-serif text-lg font-semibold leading-6 group-hover:underline group-hover:underline-offset-4">
                     {article.title}
                 </span>
-                <span className="line-clamp-4 text-sm font-normal text-gray-500">
+                <span className="line-clamp-4 text-sm font-normal text-gray-600 font-serif f">
                     {article.subtitle}
                 </span>
                 <div className="mt-2 flex items-end justify-between">
