+++ frontend/src/main.jsx
@@ -2,6 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import { BrowserRouter } from "react-router-dom"
 import App from "@/App.jsx"
+
 import "./index.css"
 
 ReactDOM.createRoot(document.getElementById("root")).render(
