+++ frontend/vite.config.js
@@ -1,6 +1,6 @@
-import { defineConfig } from "vite"
 import path from "path"
 import react from "@vitejs/plugin-react"
+import { defineConfig } from "vite"
 
 // https://vitejs.dev/config/
 export default defineConfig({
