+++ package.json
@@ -18,5 +18,8 @@
         "concurrently": "^8.2.2",
         "prettier": "^3.2.5",
         "prettier-plugin-tailwindcss": "^0.5.12"
+    },
+    "dependencies": {
+        "@ianvs/prettier-plugin-sort-imports": "^4.2.1"
     }
 }
